refactor: Storybook에서 react-hook-form의 컨텍스트를 동일하게 제공, 확장자 ts에서 tsx변경 되면서 기존 파일 삭제
